pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff>=0.6.2",
    "flask>=3.0.3",
    "loguru>=0.7.2",
    "pytest>=8.3.2",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
